update flow intro mobile app
